requirements.txt
asyncio==3.4.3
asyncpg==0.27.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
boto3==1.24.84
botocore==1.27.84
cachetools==5.2.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
DateTime==4.7
db-dtypes==1.0.4
dbus-python==1.2.16
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
entrypoints==0.3
filelock==3.8.0
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.7.1
google-cloud-appengine-logging==1.1.4
google-cloud-audit-log==0.2.4
google-cloud-bigquery==3.3.6
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-dataproc==5.0.1
google-cloud-logging==3.2.3
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==1.1.3
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
hdfs==2.7.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.4
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
mrjob==0.6.12
mysql-connector==2.2.9
netifaces==0.10.4
numpy==1.23.4
oauthlib==3.1.0
packaging==21.3
pandas==1.5.0
pandas-gbq==0.18.0
pexpect==4.6.0
platformdirs==2.5.2
proto-plus==1.22.1
protobuf==4.21.9
psycopg2-binary==2.9.3
py4j==0.10.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
pyspark==3.3.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==5.3.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
sos==4.3
SQLAlchemy==1.4.41
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
threadpoolctl==3.1.0
Twisted==18.9.0
ubuntu-advantage-tools==27.9
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
virtualenv==20.16.5
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
